+++ ChangeLog.md
@@ -3,6 +3,12 @@ Sessions for the XP Framework ChangeLog
 
 ## ?.?.? / ????-??-??
 
+## 0.2.0 / 2018-05-29
+
+* Added `Sessions::in($paths)` method to be able to influence the session 
+  cookies' paths, which defaults to "/"
+  (@thekid)
+
 ## 0.1.0 / 2018-05-18
 
 * Merged PR #2: Tie sessions to request / response - @thekid